+++ CrazyBallz/Model/IModelBall.cs
@@ -5,13 +5,13 @@ namespace Model
 {
     public abstract class IModelBall
     {
-        public static IModelBall Instantiate(int position_x, int position_y, int radius)
+        public static IModelBall Instantiate(float position_x, float position_y, int radius)
         {
             return new ModelBall(position_x, position_y, radius);
         }
 
-        public abstract int Position_X { get; set; }
-        public abstract int Position_Y { get; set; }
+        public abstract float Position_X { get; set; }
+        public abstract float Position_Y { get; set; }
         public abstract int Radius { get; }
         public abstract string Colour { get; set; }
