+++ CrazyBallz/Model/ModelBall.cs
@@ -8,12 +8,12 @@ namespace Model
 {
     internal class ModelBall : IModelBall, INotifyPropertyChanged
     {
-        private int position_x;
-        private int position_y;
+        private float position_x;
+        private float position_y;
         private int radius;
         private string colour;
 
-        public ModelBall(int position_x, int position_y, int radius)
+        public ModelBall(float position_x, float position_y, int radius)
         {
             this.position_x = position_x;
             this.position_y = position_y;
@@ -21,8 +21,8 @@ namespace Model
             this.colour = null;
         }
 
-        public override int Position_X { get => position_x; set { position_x = value; NotifyPropertyChanged(); } }
-        public override int Position_Y { get => position_y; set { position_y = value; NotifyPropertyChanged(); } }
+        public override float Position_X { get => position_x; set { position_x = value; NotifyPropertyChanged(); } }
+        public override float Position_Y { get => position_y; set { position_y = value; NotifyPropertyChanged(); } }
         public override int Radius => radius;
         public override string Colour { get => this.colour; set => this.colour = value; }
